--- data/questions/length/questions copy.txt
@@ -1,6 +1,0 @@
-How many *golf balls* tall is the *Eiffel Tower*?
-How many *watermelons* laid end to end would it take to go the maximum distance a *Boeing 747* can travel without refueling?
-If you stacked *all pennies in circulation* on top of each other, how many *Burj Khalifas* tall would they be?
-How many *Empire State Buildings* would it take to wrap all the way around *the Earth's equator*?
-How many *wavelengths of blue light* long is *The Titanic*?
-*Saturn's distance from the sun* is how many times greater than *the max depth of the Pacific Ocean*?
